Move ThemeButton to its own file
