+++ Tireless/Auth/View/LoginView.swift
@@ -68,33 +68,6 @@ struct LoginView: View {
     }
 }
 
-struct ThemeButton: View {
-    
-    let width: CGFloat
-    let name: String
-    let action: () -> Void
-    
-    var body: some View {
-        Button {
-            action()
-        } label: {
-            Text(name)
-                .padding(.top, 20)
-                .padding(.bottom, 20)
-                .frame(minWidth: width, maxWidth: .infinity,  minHeight: 44)
-                .font(Font.title3.bold())
-                .background(RoundedRectangle(cornerRadius: 12).fill(.brown))
-                .foregroundColor(Color.white)
-                .cornerRadius(8)
-        }
-        .buttonStyle(GrowingButton())
-        .lineLimit(2)
-        .multilineTextAlignment(.center)
-        .shadow(color: Color.black.opacity(0.2), radius: 5, x: 0, y: 5)
-        .padding(.bottom, 40)
-    }
-}
-
 struct PolicyView: View {
     @State private var policyIsPresenting = false
     @State private var termsIsPresenting = false
